OK sauf Delete => Fixer les delete en cascade ....

+++ src/main/java/be/digitalcity/projetspringrest/services/ProductService.java
@@ -4,7 +4,9 @@ import be.digitalcity.projetspringrest.mappers.ProductMapper;
 import be.digitalcity.projetspringrest.models.dtos.ProductDto;
 import be.digitalcity.projetspringrest.models.entities.Product;
 import be.digitalcity.projetspringrest.models.forms.ProductForm;
+import be.digitalcity.projetspringrest.repositories.OmnithequeRepository;
 import be.digitalcity.projetspringrest.repositories.ProductRepository;
+import be.digitalcity.projetspringrest.repositories.UsersRepository;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +18,16 @@ public class ProductService {
     private final ProductMapper mapper;
     private final ProductRepository repository;
     private final OmnithequeService omnithequeService;
+    private final UsersRepository usersRepository;
+    private final OmnithequeRepository omnithequeRepository;
 
-    public ProductService(ProductMapper mapper, ProductRepository repository, OmnithequeService omnithequeService) {
+    public ProductService(ProductMapper mapper, ProductRepository repository, OmnithequeService omnithequeService, UsersRepository usersRepository, OmnithequeRepository omnithequeRepository) {
         this.mapper = mapper;
         this.repository = repository;
 
         this.omnithequeService = omnithequeService;
+        this.usersRepository = usersRepository;
+        this.omnithequeRepository = omnithequeRepository;
     }
 
     public ProductDto getOne(Long id){
@@ -33,22 +39,31 @@ public class ProductService {
         return repository.findAll().stream().map(mapper::entityToDto).toList();
     }
 
-    public ProductDto create(Authentication auth,  Long omnithequeId, ProductForm form){
-        if( omnithequeId == null || form == null ) throw new IllegalArgumentException("le formulaire et l'id de l'omnitheque ne peuvent pas être null");
+    public ProductDto create(Authentication auth, ProductForm form){
+        if( form == null ) throw new IllegalArgumentException("le formulaire et l'id de l'omnitheque ne peuvent pas être null");
         Product product = mapper.formToEntity( form );
-        return mapper.entityToDto(omnithequeService.addProduct(auth,omnithequeId, product));
+        Long omnithequeId = usersRepository.findByEmail(auth.getName()).get().getOmnitheque().getId();
+        ProductDto productDto = mapper.entityToDto(omnithequeService.addProduct(auth,omnithequeId, product));
+        productDto.setOmnithequeId((omnithequeId));
+        return productDto;
     }
 
-    public ProductDto update(Long id,ProductForm form, Authentication auth){
-        if(form == null || id == null)
+    public ProductDto update(Authentication auth, ProductForm form){
+        if(form == null )
             throw new IllegalArgumentException("le formulaire et l'id ne peuvent pas être null");
 
-        Product toUpdate = repository.findById(id).orElseThrow(()->new EntityNotFoundException("Aucun produit trouvé avec l'id {"+id+"}"));
+        Product toUpdate = repository.findById(form.getId()).orElseThrow(()->new EntityNotFoundException("Aucun produit trouvé avec l'id {"+form.getId()+"}"));
 
-        if(form.getName() != null) toUpdate.setName(form.getName());
-        if(form.getCategory() != null) toUpdate.setCategory(form.getCategory());
+        if(usersRepository.findByEmail(auth.getName()).get().getOmnitheque().getProductList().stream().anyMatch(p->p.getId().equals(form.getId())))
+            if(form.getName() != null) toUpdate.setName(form.getName());
+            if(form.getCategory() != null) toUpdate.setCategory(form.getCategory());
+            if(form.getDescription() != null) toUpdate.setDescription(form.getDescription());
+            if(form.getImage() != null) toUpdate.setImage(form.getImage());
+            if(form.getQuantity() != null) toUpdate.setQuantity(form.getQuantity());
+            toUpdate = repository.save(toUpdate);
 
-        return mapper.entityToDto(repository.save(toUpdate));
+
+        return mapper.entityToDto(toUpdate);
     }
 
     public ProductDto delete(Long id, Authentication auth){
@@ -56,7 +71,10 @@ public class ProductService {
             throw new IllegalArgumentException("l'id ne peut pas être null");
 
         Product product = repository.findById(id).orElseThrow(()->new EntityNotFoundException("Aucun produit trouvé avec l'id {"+id+"}"));
-        repository.delete(product);
+        if(usersRepository.findByEmail(auth.getName()).get().getOmnitheque().getProductList().stream().anyMatch(p->p.getId().equals(id)))
+            repository.delete(product);
+        product.setId(null);
         return mapper.entityToDto(product);
     }
 }
+//TODO: Suppression/modification d'un produit Cascade
